back-end/requirements.txt
django
djangorestframework
overpy
osmium
djangorestframework-gis
django-cors-headers
